Minor refactoring
- Merged duplicate string resources between BXHealth and BXFinance
